change config to enum, change wording
