+++ usecases/schema/class.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/sirupsen/logrus"
 	"github.com/weaviate/weaviate/adapters/repos/db/inverted/stopwords"
 	"github.com/weaviate/weaviate/entities/backup"
 	"github.com/weaviate/weaviate/entities/classcache"
@@ -80,176 +79,6 @@ func (h *Handler) GetCachedClass(ctxWithClassCache context.Context,
 	return classcache.ClassesFromContext(ctxWithClassCache, func(names ...string) (map[string]versioned.Class, error) {
 		return h.classGetter.getClasses(names)
 	}, names...)
-}
-
-const (
-	GetClassAlwaysLeader            = "always_leader"
-	GetClassAlwaysLocal             = "always_local"
-	GetClassLeaderIfVersionMismatch = "leader_if_version_mismatch"
-)
-
-type classGetter interface {
-	getClasses(names []string) (map[string]versioned.Class, error)
-}
-
-func NewClassGetter(getClassMethod string, schemaParser *Parser, schemaManager SchemaManager, schemaReader SchemaReader, logger logrus.FieldLogger) (classGetter, error) {
-	switch strings.ToLower(getClassMethod) {
-	case GetClassAlwaysLeader:
-		return newClassGetterFromLeader(schemaParser, schemaManager, logger), nil
-	case GetClassAlwaysLocal:
-		return newClassGetterFromLocal(schemaParser, schemaManager, schemaReader, logger), nil
-	case GetClassLeaderIfVersionMismatch:
-		return newClassGetterFromLeaderIfVersionMismatch(schemaReader, logger), nil
-	default:
-		return nil, fmt.Errorf("unknown class getter method: %s", getClassMethod)
-	}
-}
-
-type classGetterFromLeader struct {
-	parser        *Parser
-	schemaManager SchemaManager
-	logger        logrus.FieldLogger
-}
-
-func newClassGetterFromLeader(parser *Parser, schemaManager SchemaManager, logger logrus.FieldLogger) *classGetterFromLeader {
-	return &classGetterFromLeader{
-		parser:        parser,
-		schemaManager: schemaManager,
-		logger:        logger,
-	}
-}
-
-type classGetterFromLocal struct {
-	parser        *Parser
-	schemaManager SchemaManager
-	schemaReader  SchemaReader
-	logger        logrus.FieldLogger
-}
-
-func newClassGetterFromLocal(parser *Parser, schemaManager SchemaManager, schemaReader SchemaReader, logger logrus.FieldLogger) *classGetterFromLocal {
-	return &classGetterFromLocal{
-		parser:        parser,
-		schemaManager: schemaManager,
-		schemaReader:  schemaReader,
-		logger:        logger,
-	}
-}
-
-type classGetterFromLeaderIfVersionMismatch struct {
-	schemaReader SchemaReader
-	logger       logrus.FieldLogger
-}
-
-func newClassGetterFromLeaderIfVersionMismatch(schemaReader SchemaReader, logger logrus.FieldLogger) *classGetterFromLeaderIfVersionMismatch {
-	return &classGetterFromLeaderIfVersionMismatch{
-		schemaReader: schemaReader,
-		logger:       logger,
-	}
-}
-
-func (cg *classGetterFromLeader) getClasses(names []string) (map[string]versioned.Class, error) {
-	vclasses, err := cg.schemaManager.QueryReadOnlyClasses(names...)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(vclasses) == 0 {
-		return nil, nil
-	}
-
-	for _, vclass := range vclasses {
-		if err := cg.parser.ParseClass(vclass.Class); err != nil {
-			// remove invalid classes
-			cg.logger.WithFields(logrus.Fields{
-				"Class": vclass.Class.Class,
-				"Error": err,
-			}).Warn("parsing class error")
-			delete(vclasses, vclass.Class.Class)
-			continue
-		}
-	}
-
-	return vclasses, nil
-}
-
-func (cg *classGetterFromLocal) getClasses(names []string) (map[string]versioned.Class, error) {
-	classVersions, err := cg.schemaManager.QueryClassVersions(names...)
-	if err != nil {
-		return nil, err
-	}
-	versionedClassesToReturn := map[string]versioned.Class{}
-	versionedClassesToQueryFromLeader := []string{}
-	for _, name := range names {
-		localVclass, err := cg.schemaReader.ReadOnlyVersionedClass(name)
-		leaderClassVersion, ok := classVersions[name]
-		// < leaderClassVersion instead of != because there is some chance that the local version
-		// could be ahead of the version returned by the leader if the response from the leader was
-		// delayed and i don't think it would be helpful to query the leader again in that case as
-		// it would likely return a version that is at least as large as the local version.
-		if err != nil || !ok || localVclass.Version < leaderClassVersion {
-			versionedClassesToQueryFromLeader = append(versionedClassesToQueryFromLeader, name)
-			continue
-		}
-		versionedClassesToReturn[name] = localVclass
-	}
-	if len(versionedClassesToQueryFromLeader) == 0 {
-		return versionedClassesToReturn, nil
-	}
-
-	versionedClassesFromLeader, err := cg.schemaManager.QueryReadOnlyClasses(versionedClassesToQueryFromLeader...)
-	if err != nil || len(versionedClassesFromLeader) == 0 {
-		cg.logger.WithFields(logrus.Fields{
-			"classes":    versionedClassesToQueryFromLeader,
-			"error":      err,
-			"suggestion": "look for network issues between this node and the leader",
-		}).Warn("unable to query classes from leader")
-		// return as many classes as we could get (to match previous behavior of the caller)
-		return versionedClassesToReturn, err
-	}
-
-	for _, vclass := range versionedClassesFromLeader {
-		if err := cg.parser.ParseClass(vclass.Class); err != nil {
-			// silently remove invalid classes to match previous behavior
-			cg.logger.WithFields(logrus.Fields{
-				"Class": vclass.Class.Class,
-				"Error": err,
-			}).Warn("parsing class error")
-			delete(versionedClassesFromLeader, vclass.Class.Class)
-			continue
-		}
-		versionedClassesToReturn[vclass.Class.Class] = vclass
-	}
-
-	return versionedClassesToReturn, nil
-}
-
-func (cg *classGetterFromLeaderIfVersionMismatch) getClasses(names []string) (map[string]versioned.Class, error) {
-	vclasses := map[string]versioned.Class{}
-	for _, name := range names {
-		vc, err := cg.schemaReader.ReadOnlyVersionedClass(name)
-		if err != nil {
-			cg.logger.WithFields(logrus.Fields{
-				"Class": vc.Class.Class,
-				"Error": err,
-			}).Warn("parsing local class error")
-			continue
-		}
-		vclasses[name] = vc
-	}
-	if len(vclasses) < len(names) {
-		missingClasses := []string{}
-		for _, name := range names {
-			if _, ok := vclasses[name]; !ok {
-				missingClasses = append(missingClasses, name)
-			}
-		}
-		cg.logger.WithFields(logrus.Fields{
-			"missing": missingClasses,
-			"suggestion": "if this node is too slow to pick up new classes from the leader, " +
-				"consider changing the GET_CLASS_METHOD config to `always_leader`",
-		}).Warn("not all classes found locally")
-	}
-	return vclasses, nil
 }
 
 // AddClass to the schema
